Audio update, see jump
 - add rtmidi dependency and create MIDIEngine struct
 - support event handlers in MIDIEngine
 - audio_time_t is now a long double in seconds
 - define ZG_PI as acos(-1)
 - Add Oscillator ISoundNode implementation with various wave types

+++ include/zg/audio/audio_time_t.hpp
@@ -2,5 +2,5 @@
 #include <stdint.h>
 namespace zg::audio
 {
-	typedef uint64_t audio_time_t;
+	typedef long double audio_time_t;
 }
